d/rules.d: rtla: Update Makefile for upstream restructuring
In upstream commit 01474dc706ca ("tools/rtla: Use tools/build makefiles
to build rtla") the ``Makefile`` was completely restructered to make
use of the ``tools/build`` infrastructure.

It now assumes that it is built in-tree, so copy the parts it actually
uses into the place where it now expects them (locations are relative).

Note that the upstream commit message has the following line:
"The tarball method was set as deprecated."