[rt] pinctrl: renesas: rzg2l: use spin_{lock,unlock}_irq{save,restore}
Since 6.9-rc7 there has been a build failure on arm64, but not on amd64,
and this happened during the RT part of the build.

The patch that fixes that is now added to renesas-pinctrl-fixes-for-v6.10,
which should find its way to 6.9 after being merged by Linus.

Add the patch to ``debian/patches-rt`` as that process may take a while
and it has been added to patches-6.10-rc1-rt1.tar.xz.
*If* there is another update for the 6.9 rt patch set, then it will
likely include this patch, so 'backport' the patch from 6.10-rc1-rt1.

Link: https://lore.kernel.org/linux-rt-users/131999629.KQPSlr0Zke@bagend/
Link: https://lore.kernel.org/linux-renesas-soc/20240522055421.2842689-1-claudiu.beznea.uj@bp.renesas.com/
Link: https://git.kernel.org/pub/scm/linux/kernel/git/geert/renesas-drivers.git/log/?h=renesas-pinctrl-fixes-for-v6.10