[arm64] Change COMMON_CLK_MT8188, COMMON_CLK_MT8365 to be modules
The MediaTek SoC changes merged in a previous commit set the two clock
drivers for MT8188 and MT8365 SoCs to be built-in. These can be built as
modules. The original author didn't mention that they needed these
specifically to be built-in. Assume it was out of habit and turn the
two drivers into modules.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>