[arm64] Enable more clock drivers for various MediaTek SoCs
The MediaTek SoC changes merged in a previous commit enable all MediaTek
related pin controller drivers, but only a few clock drivers to match.
Considering the amount of other drivers enabled for related devices,
it's possible that these clock drivers being missing is the biggest
problem for some of these devices.

Enable the rest of the MediaTek clock drivers. The MT6765, MT6797 and
MT7981 ones can't be built as modules, so make them built-in but build
others as modules.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>