d/b/buildcheck.py: Restore config checks on kernels to be signed
We must only sign kernel images that honour the Secure Boot setting.
gencontrol_signed.py used to check the kernel config file to ensure
that the required options (CONFIG_EFI_STUB and
CONFIG_LOCK_DOWN_IN_EFI_SECURE_BOOT) were enabled, but this was lost
when its functionality was moved into gencontrol.py.

We must also list any certificates trusted by the kernel in files.json
in the signing template.  We used to generate this from the
CONFIG_SYSTEM_TRUSTED_KEYS variable in kernel config, but that is now
set to an empty string and gencontrol.py assumes that is the case.

Restore the config checks in buildcheck.py, in the setup phase.

Fixes: 6601ee566678 ("Don't trust Secure Boot key any longer")
Fixes: 113582f8c61a ("Use early generated signing template info")