debian/rules: sort control.md5sums to improve reproducibility
The order of entries in generated `control.md5sum` can vary
depending on the system on which the command `make -f debian/rules
debian/control-real` is run, making efforts to quickly verify
that no changes were made by running this command difficult.

Add a sort step to the md5sum command so that entries are in a more
controlled order, improving reproducibility.

Signed-off-by: Martyn Welch <martyn.welch@collabora.com>
Signed-off-by: Dylan Aïssi <dylan.aissi@collabora.com>