[i386] Stop building kernel packages
The upstream Linux/x86 maintainers no longer spend much effort on
32-bit (i386) configurations, while the compat mode of 64-bit (amd64)
kernels is well-tested.

As discussed previously with the release team, i386 user-space should
continue to be supported through chroots or multiarch on amd64 hosts,
but i386 kernel packages should not be provided for future releases.

This removes all configuration for i386 kernel packages, while leaving
user-space packages untouched.

In debian/templates/image.bug/include-model, it also removes the last
reference to ia64, which was missed in the earlier removal.