[arm64,armhf] udeb: Add all watchdog drivers to kernel-image
Currently we include a selected few watchdog drivers in kernel-image
that we know are needed: either to pet a watchdog that is started by
the boot loader on a particular device, or to reboot at the end of
installation.

A recent bug report pointed out that some i.MX6 (armhf) systems need
the imx2_wdt driver because the boot loader starts the watchdog.

Rather than handling these cases one by one, include all watchdog
drivers in kernel-image.

Closes: #1081550